remove writing methods, clean up
The methods for writing flow profiles and conditions files have been
removed from the FlowExperiment object and moved to the Writers module.
The example was updated accordingly. The Utils module was cleaned up and
general linting was performed using mypy, formatting with Black.
The text encoding for .nfp files has been changed from utf-8 to cp1252.